requirements.txt
Django
djangorestframework
drf-spectacular
django-cors-headers
gunicorn